fix: code optimization and small changes
