+++ app.php
@@ -5,7 +5,7 @@ declare(strict_types= 1);
 use App\Entity\User\User;
 use App\Entity\Person\Name;
 use App\Helper\ExceptionHandler;
-use App\Repository\InMemoryUserRepository;
+use App\Repository\UserRepository\InMemoryUserRepository;
 
 require_once 'vendor/autoload.php';
 require_once 'functions.php';
@@ -18,7 +18,7 @@ $repository = new InMemoryUserRepository();
 try {
     $repository->save($user);
 
-    echo $repository->getById(1);
+    echo $repository->getUser(1);
 } catch (\Throwable $e) {
     (new ExceptionHandler())->handle($e);
 }
