+++ src/Exception/BaseAppException.php
@@ -9,7 +9,7 @@ class BaseAppException extends Exception
 {
     private array $data = [];
 
-    public function __construct($message = "", $code = 0, array $data = [])
+    public function __construct($message = "", array $data = [], $code = 500)
     {
         parent::__construct($message, $code);
         $this->data = $data;
