+++ src/Exception/UserNotFoundException.php
@@ -5,8 +5,8 @@ use App\Exception\BaseAppException;
 
 class UserNotFoundException extends BaseAppException
 {
-    public function __construct($message = "Пользователь не найден", $code = 404, $data = [])
+    public function __construct($message = "Пользователь не найден", $data = [], $code = 404)
     {
-        parent::__construct($message, $code, $data);
+        parent::__construct($message, $data, $code);
     }
 }
