+++ src/Exception/WrapperException.php
@@ -8,6 +8,6 @@ class WrapperException extends BaseAppException
 {
     public function __construct(Exception $exception) 
     {
-        parent::__construct($exception->getMessage(), $exception->getCode(), []);
+        parent::__construct($exception->getMessage(), [], $exception->getCode());
     }
 }
